+++ CHANGELOG.md
@@ -11,6 +11,11 @@ All notable changes to this project will be documented in this file.
 _This chance is used when variations can't be determined programmatically. 
 E.g. The Wither variants and Screaming goats._
 
+
+### Changed
+
+- Changed `PositionUtil` to improve safe position finding for teleportation. 
+
 ### Fixed
 
 - Fixed name for White Splotched Rabbit head.